refactor: BarChartTemplate, PieChartTemplate  추상화 , title 컴포넌트 분리
