+++ src/components/common/Title.tsx
@@ -1,0 +1,3 @@
+export const Title = ({ title }: { title: string }) => {
+  return <h3 className="text-lg font-bold">{title}</h3>;
+};
